- 64 bit updates (3): SDK and main C::B application - updated build batch files - undefine CB_PRECOMP if NOPCH is defined (to easier find NOPCH fixes)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@7827 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/projectbuildtarget.cpp
@@ -166,7 +166,7 @@ ProjectFile* ProjectBuildTarget::GetFile(int index)
         }
     }
 
-    if (index < 0 || index >= m_FileArray.GetCount())
+    if (index < 0 || static_cast<size_t>(index) >= m_FileArray.GetCount())
         return NULL;
 
     return m_FileArray.Item(index);
